requirements.txt
pillow
google-generativeai
pandas
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
google-api-core
